requirements.txt
numpy==1.22.3
torch==1.13.1 
torchaudio==0.13.1
fire
PyYAML==6.0.2
joblib==1.4.0
scikit-learn==1.3.2
tqdm
librosa==0.8.0
transformers==4.40.1
fairseq==0.12.2